Move new data-replication tests to old files
