+++ tests/integration/high_availability/test_replication_reelection.py
@@ -2,73 +2,97 @@
 # See LICENSE file for licensing details.
 
 import logging
-import time
 
+import jubilant_backports
 import lightkube
+import pytest
+from jubilant_backports import Juju
 from lightkube.resources.core_v1 import Pod
-from pytest_operator.plugin import OpsTest
 
-from ..helpers import (
-    get_primary_unit,
+from ..helpers import generate_random_string
+from .high_availability_helpers_new import (
+    CHARM_METADATA,
+    check_mysql_instances_online,
+    check_mysql_units_writes_increment,
+    get_mysql_instance_label,
+    get_mysql_primary_unit,
+    insert_mysql_test_data,
+    remove_mysql_test_data,
+    wait_for_apps_status,
 )
-from .high_availability_helpers import (
-    clean_up_database_and_table,
-    ensure_all_units_continuous_writes_incrementing,
-    ensure_n_online_mysql_members,
-    get_application_name,
-    insert_data_into_mysql_and_validate_replication,
-)
 
-logger = logging.getLogger(__name__)
+MYSQL_APP_NAME = "mysql-k8s"
+MYSQL_TEST_APP_NAME = "mysql-test-app"
 
-TIMEOUT = 15 * 60
+MINUTE_SECS = 60
+
+logging.getLogger("jubilant.wait").setLevel(logging.WARNING)
 
 
-async def test_kill_primary_check_reelection(
-    ops_test: OpsTest, highly_available_cluster, continuous_writes, credentials
-) -> None:
-    """Test to kill the primary under load and ensure re-election of primary."""
-    mysql_application_name = get_application_name(ops_test, "mysql")
-    assert mysql_application_name, "mysql application not found"
-
-    await ensure_all_units_continuous_writes_incrementing(ops_test, credentials=credentials)
-
-    mysql_unit = ops_test.model.applications[mysql_application_name].units[0]
-    primary = await get_primary_unit(ops_test, mysql_unit, mysql_application_name)
-    primary_name = primary.name
-
-    # kill the primary pod
-    client = lightkube.Client()
-    client.delete(Pod, primary.name.replace("/", "-"), namespace=ops_test.model.info.name)
-
-    time.sleep(60)
-
-    async with ops_test.fast_forward("60s"):
-        # wait for model to stabilize, k8s will re-create the killed pod
-        await ops_test.model.wait_for_idle(
-            apps=[mysql_application_name],
-            status="active",
-            raise_on_blocked=True,
-            timeout=TIMEOUT,
-            idle_period=30,
-        )
-
-        # ensure a new primary was elected
-        mysql_unit = ops_test.model.applications[mysql_application_name].units[0]
-        new_primary = await get_primary_unit(ops_test, mysql_unit, mysql_application_name)
-        new_primary_name = new_primary.name
-
-        assert primary_name != new_primary_name
-
-        # wait (and retry) until the killed pod is back online in the mysql cluster
-        assert await ensure_n_online_mysql_members(ops_test, 3), (
-            "Old primary has not come back online after being killed"
-        )
-
-    await ensure_all_units_continuous_writes_incrementing(ops_test, credentials=credentials)
-
-    database_name, table_name = "test-kill-primary-check-reelection", "data"
-    await insert_data_into_mysql_and_validate_replication(
-        ops_test, database_name, table_name, credentials
+@pytest.mark.abort_on_fail
+def test_deploy_highly_available_cluster(juju: Juju, charm: str) -> None:
+    """Simple test to ensure that the MySQL and application charms get deployed."""
+    logging.info("Deploying MySQL cluster")
+    juju.deploy(
+        charm=charm,
+        app=MYSQL_APP_NAME,
+        base="ubuntu@22.04",
+        config={"profile": "testing"},
+        resources={"mysql-image": CHARM_METADATA["resources"]["mysql-image"]["upstream-source"]},
+        num_units=3,
     )
-    await clean_up_database_and_table(ops_test, database_name, table_name, credentials)
+    juju.deploy(
+        charm=MYSQL_TEST_APP_NAME,
+        app=MYSQL_TEST_APP_NAME,
+        base="ubuntu@22.04",
+        channel="latest/edge",
+        config={"sleep_interval": 300},
+        num_units=1,
+    )
+
+    juju.integrate(
+        f"{MYSQL_APP_NAME}:database",
+        f"{MYSQL_TEST_APP_NAME}:database",
+    )
+
+    logging.info("Wait for applications to become active")
+    juju.wait(
+        ready=wait_for_apps_status(
+            jubilant_backports.all_active, MYSQL_APP_NAME, MYSQL_TEST_APP_NAME
+        ),
+        error=jubilant_backports.any_blocked,
+        timeout=20 * MINUTE_SECS,
+    )
+
+
+@pytest.mark.abort_on_fail
+def test_kill_primary_check_reelection(juju: Juju) -> None:
+    """Confirm that a new primary is elected when the current primary is tear down."""
+    check_mysql_units_writes_increment(juju, MYSQL_APP_NAME)
+
+    mysql_old_primary = get_mysql_primary_unit(juju, MYSQL_APP_NAME)
+    mysql_old_primary_label = get_mysql_instance_label(mysql_old_primary)
+
+    logging.info("Killing the primary pod")
+    client = lightkube.Client()
+    client.delete(Pod, mysql_old_primary_label, namespace=juju.model)
+
+    juju.wait(
+        ready=wait_for_apps_status(jubilant_backports.all_active, MYSQL_APP_NAME),
+        error=jubilant_backports.any_blocked,
+        timeout=20 * MINUTE_SECS,
+    )
+
+    # Confirm that the new primary unit is different
+    mysql_new_primary = get_mysql_primary_unit(juju, MYSQL_APP_NAME)
+    assert mysql_new_primary != mysql_old_primary, "Primary has not changed"
+
+    # Retry until the killed pod is back online in the mysql cluster
+    assert check_mysql_instances_online(juju, MYSQL_APP_NAME)
+
+    table_name = "data"
+    table_value = generate_random_string(255)
+
+    check_mysql_units_writes_increment(juju, MYSQL_APP_NAME)
+    insert_mysql_test_data(juju, MYSQL_APP_NAME, table_name, table_value)
+    remove_mysql_test_data(juju, MYSQL_APP_NAME, table_name)
