--- tests/spread/test_replication_reelection_new.py/task.yaml
@@ -1,7 +1,0 @@
-summary: test_replication_reelection_new.py
-environment:
-  TEST_MODULE: high_availability/test_replication_reelection_new.py
-execute: |
-  tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
-artifacts:
-  - allure-results
